backend/sentiment_analysis_backend/requirements.txt
asgiref==3.4.1
click==8.0.4
Django==3.2.13
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
nltk==3.6.7
numpy==1.19.5
Pillow==8.4.0
pytesseract==0.3.7
pytz==2022.1
regex==2022.3.15
scikit-learn==0.24.2
scipy==1.5.4
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.1.1
zipp==3.6.0
